Refactor packages and add `component` parameter to the image Composable functions
