+++ landscapist/src/main/kotlin/com/skydoves/landscapist/plugins/ImagePlugin.kt
@@ -13,15 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.skydoves.landscapist
+package com.skydoves.landscapist.plugins
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.Immutable
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.graphics.painter.Painter
 
 /**
- * A pluggable compose interface that will be executed for image loadings.
+ * A pluggable compose interface that will be executed for loading images.
  */
+@Immutable
 public sealed interface ImagePlugin {
 
   /**
